requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
langchain-groq
langchain-huggingface
pypdf
chromadb